// src/interfaces/TextProps/TextProps.tsx
export interface TextProps {
	text?: string;
    color?: string;
    fontSize?: string;
    fontWeight?: string;
    fontStyle?: string;
    textTransform?: string;
    fontFamily?: string;
    whiteSpace?: string;
    secondColor?: string;
	secondText?: string;
    secondfontFamily?: string;
    textAlign?: string;
    overlaid?: boolean;
    bottom?: string;
    width?: string;
    inline?: boolean;
}